requirements.txt
aiohttp==3.6.2
amqp==2.6.0
apipkg==1.5
appdirs==1.4.4
arrow==0.15.7
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
bandit==1.6.2
billiard==3.6.3.0
black==19.10b0
blessed==1.17.8
blis==0.4.1
CacheControl==0.12.6
catalogue==1.0.0
celery==4.4.6
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
coverage==5.1
cymem==2.0.3
distlib==0.3.0
distro==1.4.0
Django==3.0.8
django-annoying==0.10.6
django-appconf==1.0.4
django-archive==0.1.6
django-autoslug-iplweb==1.9.4.dev0
django-cleanup==5.0.0
django-compressor==2.4
django-crispy-forms==1.9.1
django-hitcount==1.3.2
django-markdownx==3.0.1
django-material==1.6.7
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-picklefield==3.0.1
django-q==1.3.2
django-q-sentry==0.1.4
django-taggit==1.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.10.3
docopt==0.6.2
en-core-web-sm==2.3.1
execnet==1.7.1
flake8==3.8.3
future==0.18.2
fuzzywuzzy==0.17.0
gevent==20.6.2
gitdb==4.0.5
GitPython==3.1.3
greenlet==0.4.16
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
inflect==4.1.0
ipaddr==2.2.0
isort==4.3.21
joblib==0.16.0
jprq==1.0.4
jsonfield==3.1.0
kombu==4.6.11
lazy-object-proxy==1.4.3
lockfile==0.12.2
loguru==0.5.3
Markdown==3.2.2
martor==1.5.0
mccabe==0.6.1
more-itertools==8.4.0
msgpack==0.6.2
multidict==4.7.6
murmurhash==1.0.2
nltk==3.5
numpy==1.19.0
packaging==20.4
pandas==1.0.5
pathspec==0.8.0
pbr==5.4.5
pep517==0.8.2
Pillow==7.1.2
pipreqs==0.4.10
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
progress==1.5
pusher-chatkit-server==1.0.4
py==1.8.2
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.3
pylint-runner==0.6.0
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
pytest-forked==1.1.3
pytest-xdist==1.32.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytoml==0.1.21
pytz==2019.3
PyYAML==5.3.1
rcssmin==1.0.6
redis==3.5.3
regex==2020.6.8
requests==2.24.0
retrying==1.3.3
rjsmin==1.1.0
rope==0.17.0
sentry-sdk==0.16.5
six==1.15.0
smmap==3.0.4
sorl-thumbnail==12.6.3
spacy==2.3.2
sqlparse==0.3.1
srsly==1.0.2
stevedore==2.0.0
swapper==1.1.2.post1
textblob==0.15.3
thinc==7.4.1
toml==0.10.1
tqdm==4.48.0
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
wasabi==0.8.0
wcwidth==0.2.4
webencodings==0.5.1
websockets==8.1
wincertstore==0.2
word2number==1.1
wrapt==1.12.1
yarg==0.1.9
yarl==1.6.2
zope.event==4.4
zope.interface==5.1.0
